RJ AND EJ DONE

+++ Projetos/Projeto-Particula/src/utils/singly_linked_list.c
@@ -81,7 +81,7 @@ bool list_is_empty(List list)
  * @param list The linked list.
  * @return size_t The number of elements in the list.
  */
-size_t list_size(List list)
+int list_size(List list)
 {
     return list->size;
 }
